Risky cards DRY
