+++ src/components/cards/proposal/RiskyInvestor/styled.tsx
@@ -1,51 +1,12 @@
-import { ReactNode, FC, useMemo } from "react"
+import { FC, useMemo } from "react"
 import styled, { css } from "styled-components"
 import { motion, AnimatePresence } from "framer-motion"
 
-import { Flex, GradientBorder } from "theme"
-import Amount from "components/Amount"
-
-import checkGreenIcon from "assets/icons/green-check.svg"
+import { Flex } from "theme"
 import starIcon from "assets/icons/star.svg"
 import starDarkIcon from "assets/icons/star-dark.svg"
 
 const Styled = {
-  Container: styled(GradientBorder)`
-    width: 100%;
-    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1);
-    border-radius: 16px;
-    flex-direction: column;
-    margin-bottom: 18px;
-
-    &:after {
-      background: #181e2c;
-    }
-  `,
-  Head: styled(Flex)`
-    width: 100%;
-    justify-content: space-between;
-    padding: 8px 14px 7px 16px;
-    border-bottom: 1px solid #1d2635;
-    position: relative;
-  `,
-  Body: styled.div`
-    width: 100%;
-    padding: 12px 14px 16px 16px;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-template-rows: repeat(3, 1fr);
-    gap: 16px 8px;
-  `,
-  Title: styled.div`
-    font-family: "Gilroy";
-    font-style: normal;
-    font-weight: 600;
-    font-size: 14px;
-    line-height: 100%;
-    color: #e4f2ff;
-    margin: 0 4px;
-    transform: translateY(2px);
-  `,
   Status: styled.div<{ active?: boolean }>`
     padding: 5px 6px;
     border-radius: 36px;
@@ -114,52 +75,6 @@ const Styled = {
 }
 
 export default Styled
-
-// Body item
-export const BodyItemStyled = {
-  Container: styled(Flex)`
-    width: 100%;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: flex-start;
-  `,
-  Label: styled(Flex)`
-    min-height: 16px;
-    align-items: center;
-    margin-bottom: 2px;
-    font-family: "Gilroy";
-    font-style: normal;
-    font-weight: 400;
-    font-size: 11px;
-    line-height: 100%;
-    letter-spacing: 0.03em;
-    color: #616d8b;
-  `,
-}
-
-interface IBodyItemProps {
-  label: ReactNode
-  amount: ReactNode
-  symbol?: ReactNode
-  fz?: string
-  completed?: boolean
-}
-
-export const BodyItem: FC<IBodyItemProps> = ({
-  label,
-  amount,
-  symbol,
-  fz,
-  completed = false,
-}) => (
-  <BodyItemStyled.Container>
-    <BodyItemStyled.Label>
-      {label}
-      {completed && <img src={checkGreenIcon} />}
-    </BodyItemStyled.Label>
-    <Amount value={amount} symbol={symbol} fz={fz} />
-  </BodyItemStyled.Container>
-)
 
 const LineBase = css`
   height: 2px;
